+++ app/src/main/java/dev/bartuzen/qbitcontroller/ui/torrentlist/TorrentListFragment.kt
@@ -440,9 +440,9 @@ class TorrentListFragment() : Fragment(R.layout.fragment_torrent_list) {
 
             binding.textSpeed.text = getString(
                 R.string.torrent_list_speed_format,
-                formatBytesPerSecond(requireContext(), mainData.serverState.uploadSpeed.toLong()),
+                formatBytesPerSecond(requireContext(), mainData.serverState.uploadSpeed),
                 formatBytes(requireContext(), mainData.serverState.uploadSession),
-                formatBytesPerSecond(requireContext(), mainData.serverState.downloadSpeed.toLong()),
+                formatBytesPerSecond(requireContext(), mainData.serverState.downloadSpeed),
                 formatBytes(requireContext(), mainData.serverState.downloadSession)
             )
